+++ feature/navigation/drawer/dropdown/src/debug/kotlin/net/thunderbird/feature/navigation/drawer/dropdown/ui/folder/FolderListPreview.kt
@@ -4,6 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
 import kotlinx.collections.immutable.persistentListOf
+import net.thunderbird.feature.navigation.drawer.dropdown.domain.entity.TreeFolder
 import net.thunderbird.feature.navigation.drawer.dropdown.ui.FakeData.DISPLAY_FOLDER
 import net.thunderbird.feature.navigation.drawer.dropdown.ui.FakeData.UNIFIED_FOLDER
 
@@ -12,8 +13,8 @@ import net.thunderbird.feature.navigation.drawer.dropdown.ui.FakeData.UNIFIED_FO
 internal fun FolderListPreview() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
-                DISPLAY_FOLDER,
+            rootFolder = TreeFolder.createFromFolders(
+                persistentListOf(DISPLAY_FOLDER),
             ),
             selectedFolder = null,
             onFolderClick = {},
@@ -27,8 +28,8 @@ internal fun FolderListPreview() {
 internal fun FolderListPreviewSelected() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
-                DISPLAY_FOLDER,
+            rootFolder = TreeFolder.createFromFolders(
+                persistentListOf(DISPLAY_FOLDER),
             ),
             selectedFolder = DISPLAY_FOLDER,
             onFolderClick = {},
@@ -42,9 +43,11 @@ internal fun FolderListPreviewSelected() {
 internal fun FolderListWithUnifiedFolderPreview() {
     PreviewWithTheme {
         FolderList(
-            folders = persistentListOf(
-                UNIFIED_FOLDER,
-                DISPLAY_FOLDER,
+            rootFolder = TreeFolder.createFromFolders(
+                persistentListOf(
+                    UNIFIED_FOLDER,
+                    DISPLAY_FOLDER,
+                ),
             ),
             selectedFolder = DISPLAY_FOLDER,
             onFolderClick = {},
