+++ feature/navigation/drawer/dropdown/src/main/kotlin/net/thunderbird/feature/navigation/drawer/dropdown/NavigationDrawerModule.kt
@@ -6,6 +6,7 @@ import net.thunderbird.feature.navigation.drawer.dropdown.domain.DomainContract.
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.GetDisplayAccounts
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.GetDisplayFoldersForAccount
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.GetDrawerConfig
+import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.GetTreeFolders
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.SaveDrawerConfig
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.SyncAccount
 import net.thunderbird.feature.navigation.drawer.dropdown.domain.usecase.SyncAllAccounts
@@ -48,6 +49,10 @@ val navigationDropDownDrawerModule: Module = module {
         )
     }
 
+    single<UseCase.GetTreeFolders> {
+        GetTreeFolders()
+    }
+
     single<UseCase.SyncAccount> {
         SyncAccount(
             accountManager = get(),
@@ -67,6 +72,7 @@ val navigationDropDownDrawerModule: Module = module {
             saveDrawerConfig = get(),
             getDisplayAccounts = get(),
             getDisplayFoldersForAccount = get(),
+            getTreeFolders = get(),
             syncAccount = get(),
             syncAllAccounts = get(),
         )
